requirements.txt
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0